+++ sklearn/cluster/tests/test_bicluster.py
@@ -13,11 +13,11 @@ from sklearn.utils.testing import SkipTest
 
 from sklearn.base import BaseEstimator, BiclusterMixin
 
-from sklearn.cluster.bicluster import SpectralCoclustering
-from sklearn.cluster.bicluster import SpectralBiclustering
-from sklearn.cluster.bicluster import _scale_normalize
-from sklearn.cluster.bicluster import _bistochastic_normalize
-from sklearn.cluster.bicluster import _log_normalize
+from sklearn.cluster import SpectralCoclustering
+from sklearn.cluster import SpectralBiclustering
+from sklearn.cluster._bicluster import _scale_normalize
+from sklearn.cluster._bicluster import _bistochastic_normalize
+from sklearn.cluster._bicluster import _log_normalize
 
 from sklearn.metrics import (consensus_score, v_measure_score)
