+++ sklearn/cluster/tests/test_dbscan.py
@@ -13,8 +13,8 @@ import pytest
 
 from sklearn.utils.testing import assert_array_equal
 from sklearn.neighbors import NearestNeighbors
-from sklearn.cluster.dbscan_ import DBSCAN
-from sklearn.cluster.dbscan_ import dbscan
+from sklearn.cluster import DBSCAN
+from sklearn.cluster import dbscan
 from sklearn.cluster.tests.common import generate_clustered_data
 from sklearn.metrics.pairwise import pairwise_distances
